+++ gcalendar/calendar_tools.py
@@ -18,6 +18,12 @@ from googleapiclient.discovery import build
 
 from auth.service_decorator import require_google_service
 from core.utils import handle_http_errors, StringList
+from gcalendar.calendar_helpers import (
+    _format_attachment_details,
+    _format_attendee_details,
+    _format_person,
+    _get_meeting_link,
+)
 
 from mcp.types import ToolAnnotations
 
@@ -206,119 +212,6 @@ def _preserve_existing_fields(
         elif new_value is not None:
             event_body[field_name] = new_value
 
-
-def _get_meeting_link(item: Dict[str, Any]) -> str:
-    """Extract video meeting link from event conference data or hangoutLink."""
-    conference_data = item.get("conferenceData")
-    if conference_data and "entryPoints" in conference_data:
-        for entry_point in conference_data["entryPoints"]:
-            if entry_point.get("entryPointType") == "video":
-                uri = entry_point.get("uri", "")
-                if uri:
-                    return uri
-    hangout_link = item.get("hangoutLink", "")
-    if hangout_link:
-        return hangout_link
-    return ""
-
-
-def _format_attendee_details(
-    attendees: List[Dict[str, Any]], indent: str = "  "
-) -> str:
-    """
-      Format attendee details including response status, organizer, and optional flags.
-
-      Example output format:
-      "  user@example.com: accepted
-    manager@example.com: declined (organizer)
-    optional-person@example.com: tentative (optional)"
-
-      Args:
-          attendees: List of attendee dictionaries from Google Calendar API
-          indent: Indentation to use for newline-separated attendees (default: "  ")
-
-      Returns:
-          Formatted string with attendee details, or "None" if no attendees
-    """
-    if not attendees:
-        return "None"
-
-    attendee_details_list = []
-    for a in attendees:
-        email = a.get("email", "unknown")
-        response_status = a.get("responseStatus", "unknown")
-        optional = a.get("optional", False)
-        organizer = a.get("organizer", False)
-
-        detail_parts = [f"{email}: {response_status}"]
-        if organizer:
-            detail_parts.append("(organizer)")
-        if optional:
-            detail_parts.append("(optional)")
-
-        attendee_details_list.append(" ".join(detail_parts))
-
-    return f"\n{indent}".join(attendee_details_list)
-
-
-def _format_attachment_details(
-    attachments: List[Dict[str, Any]], indent: str = "  "
-) -> str:
-    """
-    Format attachment details including file information.
-
-
-    Args:
-        attachments: List of attachment dictionaries from Google Calendar API
-        indent: Indentation to use for newline-separated attachments (default: "  ")
-
-    Returns:
-        Formatted string with attachment details, or "None" if no attachments
-    """
-    if not attachments:
-        return "None"
-
-    attachment_details_list = []
-    for att in attachments:
-        title = att.get("title", "Untitled")
-        file_url = att.get("fileUrl", "No URL")
-        file_id = att.get("fileId", "No ID")
-        mime_type = att.get("mimeType", "Unknown")
-
-        attachment_info = (
-            f"{title}\n"
-            f"{indent}File URL: {file_url}\n"
-            f"{indent}File ID: {file_id}\n"
-            f"{indent}MIME Type: {mime_type}"
-        )
-        attachment_details_list.append(attachment_info)
-
-    return f"\n{indent}".join(attachment_details_list)
-
-
-def _format_person(person: Optional[Dict[str, Any]]) -> Optional[str]:
-    """
-    Format a Google Calendar person dict (creator or organizer) for display.
-
-    Args:
-        person: Dict with optional ``displayName`` and ``email`` keys, or ``None``.
-
-    Returns:
-        ``"Name <email>"`` when both are present, ``"Name"`` or ``"<email>"`` when
-        only one is, or ``None`` when neither is — letting callers skip the line
-        entirely rather than printing an empty placeholder.
-    """
-    if not person:
-        return None
-    name = (person.get("displayName") or "").strip()
-    email = (person.get("email") or "").strip()
-    if name and email:
-        return f"{name} <{email}>"
-    if name:
-        return name
-    if email:
-        return f"<{email}>"
-    return None
 
 
 # Helper function to ensure time strings for API calls are correctly formatted
